Moved to support .Net Core
