+++ zxcvbn-core-test/Properties/AssemblyInfo.cs
