+++ zxcvbn-core-test/passwords.lst
