+++ zxcvbn-core-test/zxcvbn-core-test.xproj.user
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="14.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <ActiveDebugProfile>Start</ActiveDebugProfile>
+  </PropertyGroup>
+</Project>
