+++ zxcvbn-core/Dictionaries/english.lst
