+++ zxcvbn-core/Matcher/SpatialMatcher.cs
@@ -1,10 +1,30 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 
 namespace Zxcvbn.Matcher
 {
+    /// <summary>
+    /// A match made with the <see cref="SpatialMatcher"/>. Contains additional information specific to spatial matches.
+    /// </summary>
+    public class SpatialMatch : Match
+    {
+        /// <summary>
+        /// The name of the keyboard layout used to make the spatial match
+        /// </summary>
+        public string Graph { get; set; }
+
+        /// <summary>
+        /// The number of shifted characters matched in the pattern (adds to entropy)
+        /// </summary>
+        public int ShiftedCount { get; set; }
+
+        /// <summary>
+        /// The number of turns made (i.e. when diretion of adjacent keys changes)
+        /// </summary>
+        public int Turns { get; set; }
+    }
+
     /// <summary>
     /// <para>A matcher that checks for keyboard layout patterns (e.g. 78523 on a keypad, or plkmn on a QWERTY keyboard).</para>
     /// <para>Has patterns for QWERTY, DVORAK, numeric keybad and mac numeric keypad</para>
@@ -12,9 +32,9 @@ namespace Zxcvbn.Matcher
     /// </summary>
     public class SpatialMatcher : IMatcher
     {
-        const string SpatialPattern = "spatial";
+        private const string SpatialPattern = "spatial";
 
-        Lazy<List<SpatialGraph>> spatialGraphs = new Lazy<List<SpatialGraph>>(() => GenerateSpatialGraphs());
+        private Lazy<List<SpatialGraph>> spatialGraphs = new Lazy<List<SpatialGraph>>(() => GenerateSpatialGraphs());
 
         /// <summary>
         /// Match the password against the known keyboard layouts
@@ -26,65 +46,6 @@ namespace Zxcvbn.Matcher
         {
             return spatialGraphs.Value.SelectMany((g) => SpatialMatch(g, password)).ToList();
         }
-
-        /// <summary>
-        /// Match the password against a single pattern
-        /// </summary>
-        /// <param name="graph">Adjacency graph for this key layout</param>
-        /// <param name="password">Password to match</param>
-        /// <returns>List of matching patterns</returns>
-        private List<Match> SpatialMatch(SpatialGraph graph, string password)
-        {
-            var matches = new List<Match>();
-
-            var i = 0;
-            while (i < password.Length - 1)
-            {
-                int turns = 0, shiftedCount = 0;
-                var lastDirection = -1;
-
-                var j = i + 1;
-                for (; j < password.Length; ++j)
-                {
-                    bool shifted;
-                    var foundDirection = graph.GetAdjacentCharDirection(password[j - 1], password[j], out shifted);
-
-                    if (foundDirection != -1)
-                    {
-                        // Spatial match continues
-                        if (shifted) shiftedCount++;
-                        if (lastDirection != foundDirection)
-                        {
-                            turns++;
-                            lastDirection = foundDirection;
-                        }
-                    }
-                    else break; // This character not a spatial match
-
-                }
-
-                // Only consider runs of greater than two
-                if (j - i > 2) 
-                {
-                    matches.Add(new SpatialMatch()
-                    {
-                        Pattern = SpatialPattern,
-                        i = i,
-                        j = j - 1,
-                        Token = password.Substring(i, j - i),
-                        Graph = graph.Name,
-                        Entropy = graph.CalculateEntropy(j - i, turns, shiftedCount),
-                        Turns = turns,
-                        ShiftedCount = shiftedCount
-                    });
-                }
-
-                i = j;
-            }
-
-            return matches;
-        }
-
 
         // In the JS version these are precomputed, but for now we'll generate them here when they are first needed.
         private static List<SpatialGraph> GenerateSpatialGraphs()
@@ -120,7 +81,6 @@ namespace Zxcvbn.Matcher
   0 .
 ";
 
-
             return new List<SpatialGraph> { new SpatialGraph("qwerty", qwerty, true),
                     new SpatialGraph("dvorak", dvorak, true),
                     new SpatialGraph("keypad", keypad, false),
@@ -128,34 +88,126 @@ namespace Zxcvbn.Matcher
                 };
         }
 
+        /// <summary>
+        /// Match the password against a single pattern
+        /// </summary>
+        /// <param name="graph">Adjacency graph for this key layout</param>
+        /// <param name="password">Password to match</param>
+        /// <returns>List of matching patterns</returns>
+        private List<Match> SpatialMatch(SpatialGraph graph, string password)
+        {
+            var matches = new List<Match>();
+
+            var i = 0;
+            while (i < password.Length - 1)
+            {
+                int turns = 0, shiftedCount = 0;
+                var lastDirection = -1;
+
+                var j = i + 1;
+                for (; j < password.Length; ++j)
+                {
+                    bool shifted;
+                    var foundDirection = graph.GetAdjacentCharDirection(password[j - 1], password[j], out shifted);
+
+                    if (foundDirection != -1)
+                    {
+                        // Spatial match continues
+                        if (shifted) shiftedCount++;
+                        if (lastDirection != foundDirection)
+                        {
+                            turns++;
+                            lastDirection = foundDirection;
+                        }
+                    }
+                    else break; // This character not a spatial match
+                }
+
+                // Only consider runs of greater than two
+                if (j - i > 2)
+                {
+                    matches.Add(new SpatialMatch()
+                    {
+                        Pattern = SpatialPattern,
+                        i = i,
+                        j = j - 1,
+                        Token = password.Substring(i, j - i),
+                        Graph = graph.Name,
+                        Entropy = graph.CalculateEntropy(j - i, turns, shiftedCount),
+                        Turns = turns,
+                        ShiftedCount = shiftedCount
+                    });
+                }
+
+                i = j;
+            }
+
+            return matches;
+        }
+
+        // Instances of Point or Pair in the standard library are in UI assemblies, so define our own version to reduce dependencies
+        private struct Point
+        {
+            public int x;
+            public int y;
+
+            public Point(int x, int y)
+            {
+                this.x = x;
+                this.y = y;
+            }
+
+            public override string ToString()
+            {
+                return "{" + x + ", " + y + "}";
+            }
+        }
+
         // See build_keyboard_adjacency_graph.py in zxcvbn for how these are generated
         private class SpatialGraph
         {
-            public string Name { get; private set; }
-            private Dictionary<char, List<string>> AdjacencyGraph { get; set; }
-            public int StartingPositions { get; private set; }
-            public double AverageDegree { get; private set; }
-
             public SpatialGraph(string name, string layout, bool slanted)
             {
-                this.Name = name;
-                BuildGraph(layout, slanted);                
+                Name = name;
+                BuildGraph(layout, slanted);
             }
 
+            public double AverageDegree { get; private set; }
+            public string Name { get; private set; }
+            public int StartingPositions { get; private set; }
+            private Dictionary<char, List<string>> AdjacencyGraph { get; set; }
 
             /// <summary>
-            /// Returns true when testAdjacent is in c's adjacency list
+            /// Calculate entropy for a math that was found on this adjacency graph
             /// </summary>
-            public bool IsCharAdjacent(char c, char testAdjacent)
+            public double CalculateEntropy(int matchLength, int turns, int shiftedCount)
             {
-                if (AdjacencyGraph.ContainsKey(c)) return AdjacencyGraph[c].Any(s => s.Contains(testAdjacent));
-                return false;
+                // This is an estimation of the number of patterns with length of matchLength or less with turns turns or less
+                var possibilities = Enumerable.Range(2, matchLength - 1).Sum(i =>
+                {
+                    var possible_turns = Math.Min(turns, i - 1);
+                    return Enumerable.Range(1, possible_turns).Sum(j =>
+                    {
+                        return StartingPositions * Math.Pow(AverageDegree, j) * PasswordScoring.Binomial(i - 1, j - 1);
+                    });
+                });
+
+                var entropy = Math.Log(possibilities, 2);
+
+                // Entropy increaeses for a mix of shifted and unshifted
+                if (shiftedCount > 0)
+                {
+                    var unshifted = matchLength - shiftedCount;
+                    entropy += Math.Log(Enumerable.Range(0, Math.Min(shiftedCount, unshifted) + 1).Sum(i => PasswordScoring.Binomial(matchLength, i)), 2);
+                }
+
+                return entropy;
             }
 
             /// <summary>
             /// Returns the 'direction' of the adjacent character (i.e. index in the adjacency list).
             /// If the character is not adjacent, -1 is returned
-            /// 
+            ///
             /// Uses the 'shifted' out parameter to let the caller know if the matched character is shifted
             /// </summary>
             public int GetAdjacentCharDirection(char c, char adjacent, out bool shifted)
@@ -173,26 +225,18 @@ namespace Zxcvbn.Matcher
                 return AdjacencyGraph[c].IndexOf(adjacentEntry);
             }
 
-            private Point[] GetSlantedAdjacent(Point c)
+            /// <summary>
+            /// Returns true when testAdjacent is in c's adjacency list
+            /// </summary>
+            public bool IsCharAdjacent(char c, char testAdjacent)
             {
-                int x = c.x;
-                int y = c.y;
-
-                return new Point[] { new Point(x - 1, y), new Point(x, y - 1), new Point(x + 1, y - 1), new Point(x + 1, y), new Point(x, y + 1), new Point(x - 1, y + 1) };
-            }
-
-            private Point[] GetAlignedAdjacent(Point c)
-            {
-                int x = c.x;
-                int y = c.y;
-
-                return new Point[] { new Point(x - 1, y), new Point(x - 1, y - 1), new Point(x, y - 1), new Point(x + 1, y - 1), new Point(x + 1, y), new Point(x + 1, y + 1), new Point(x, y + 1), new Point(x - 1, y + 1) };
+                if (AdjacencyGraph.ContainsKey(c)) return AdjacencyGraph[c].Any(s => s.Contains(testAdjacent));
+                return false;
             }
 
             private void BuildGraph(string layout, bool slanted)
             {
-
-                var tokens =  layout.Split((char[])null, StringSplitOptions.RemoveEmptyEntries);
+                var tokens = layout.Split((char[])null, StringSplitOptions.RemoveEmptyEntries);
                 var tokenSize = tokens[0].Length;
 
                 // Put the characters in each keyboard cell into the map agains t their coordinates
@@ -229,78 +273,26 @@ namespace Zxcvbn.Matcher
                     }
                 }
 
-
-
                 // Calculate average degree and starting positions, cf. init.coffee
                 StartingPositions = AdjacencyGraph.Count;
                 AverageDegree = AdjacencyGraph.Sum(adj => adj.Value.Where(a => a != null).Count()) * 1.0 / StartingPositions;
             }
 
-            /// <summary>
-            /// Calculate entropy for a math that was found on this adjacency graph
-            /// </summary>
-            public double CalculateEntropy(int matchLength, int turns, int shiftedCount)
+            private Point[] GetAlignedAdjacent(Point c)
             {
-                // This is an estimation of the number of patterns with length of matchLength or less with turns turns or less
-                var possibilities = Enumerable.Range(2, matchLength - 1).Sum(i =>
-                {
-                    var possible_turns = Math.Min(turns, i - 1);
-                    return Enumerable.Range(1, possible_turns).Sum(j =>
-                    {
-                        return StartingPositions * Math.Pow(AverageDegree, j) * PasswordScoring.Binomial(i - 1, j - 1);
-                    });
-                });
+                int x = c.x;
+                int y = c.y;
 
-                var entropy = Math.Log(possibilities, 2);
+                return new Point[] { new Point(x - 1, y), new Point(x - 1, y - 1), new Point(x, y - 1), new Point(x + 1, y - 1), new Point(x + 1, y), new Point(x + 1, y + 1), new Point(x, y + 1), new Point(x - 1, y + 1) };
+            }
 
-                // Entropy increaeses for a mix of shifted and unshifted
-                if (shiftedCount > 0)
-                {
-                    var unshifted = matchLength - shiftedCount;
-                    entropy += Math.Log(Enumerable.Range(0, Math.Min(shiftedCount, unshifted) + 1).Sum(i => PasswordScoring.Binomial(matchLength, i)), 2);
-                }
+            private Point[] GetSlantedAdjacent(Point c)
+            {
+                int x = c.x;
+                int y = c.y;
 
-                return entropy;
+                return new Point[] { new Point(x - 1, y), new Point(x, y - 1), new Point(x + 1, y - 1), new Point(x + 1, y), new Point(x, y + 1), new Point(x - 1, y + 1) };
             }
         }
-
-        // Instances of Point or Pair in the standard library are in UI assemblies, so define our own version to reduce dependencies
-        private struct Point
-        {
-            public int x;
-            public int y;
-
-            public Point(int x, int y)
-            {
-                this.x = x;
-                this.y = y;
-            }
-
-            public override string ToString()
-            {
-                return "{" + x + ", " + y + "}";
-            }
-        }
-    }
-
-    /// <summary>
-    /// A match made with the <see cref="SpatialMatcher"/>. Contains additional information specific to spatial matches.
-    /// </summary>
-    public class SpatialMatch : Match
-    {
-        /// <summary>
-        /// The name of the keyboard layout used to make the spatial match
-        /// </summary>
-        public string Graph { get; set; }
-
-        /// <summary>
-        /// The number of turns made (i.e. when diretion of adjacent keys changes)
-        /// </summary>
-        public int Turns { get; set; }
-
-        /// <summary>
-        /// The number of shifted characters matched in the pattern (adds to entropy)
-        /// </summary>
-        public int ShiftedCount { get; set; }
     }
 }
