+++ zxcvbn-core/scripts/build_dictionaries.py
